requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
dulwich==0.21.5
filelock==3.12.2
html5lib==1.1
idna==3.4
importlib-metadata==6.7.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.1
pexpect==4.8.0
pipenv==2023.6.18
pkginfo==1.9.6
platformdirs==3.8.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
ptyprocess==0.7.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.0.post1
soupsieve==2.4.1
tomli==2.0.1
tomlkit==0.11.8
trove-classifiers==2023.5.24
urllib3==1.26.16
virtualenv==20.23.1
virtualenv-clone==0.5.7
webencodings==0.5.1
zipp==3.15.0
